box_tree: split lib.rs into modules (types, damage, util, tree)
- Move types, damage, utilities, and core tree implementation/tests into dedicated modules.
- Keep public API stable via re-exports in lib.rs.
- No behavior changes.
