+++ understory_box_tree/README.md
@@ -86,112 +86,7 @@ Key operations:
 - `examples/visible_list.rs`: demonstrates using `intersect_rect` to compute a visible set,
   a building block for virtualization.
 
-### Minimal usage
-
-```rust
-use understory_box_tree::{Tree, LocalNode, QueryFilter};
-use kurbo::{Rect, Affine, Vec2, Point};
-
-// Build a tiny tree.
-let mut tree = Tree::new();
-
-let root = tree.insert(
-    None,
-    LocalNode { local_bounds: Rect::new(0.0, 0.0, 200.0, 200.0), ..Default::default() },
-);
-
-let child = tree.insert(
-    Some(root),
-    LocalNode { local_bounds: Rect::new(10.0, 10.0, 60.0, 60.0), ..Default::default() },
-);
-
-// Synchronize and compute damage.
-let _ = tree.commit();
-
-// Move and hit-test.
-tree.set_local_transform(child, Affine::translate(Vec2::new(10.0, 0.0)));
-let _ = tree.commit();
-
-let filter = QueryFilter { visible_only: true, pickable_only: true };
-let hit = tree.hit_test_point(Point::new(25.0, 25.0), filter).unwrap();
-assert_eq!(hit.node, child);
-```
-
-### Visible set using a viewport rectangle
-
-```rust
-use understory_box_tree::{Tree, LocalNode, QueryFilter};
-use kurbo::Rect;
-
-let mut tree = Tree::new();
-
-let root = tree.insert(
-    None,
-    LocalNode { local_bounds: Rect::new(0.0, 0.0, 1000.0, 1000.0), ..Default::default() },
-);
-
-// Insert rows.
-for i in 0..10u32 {
-    let y = i as f64 * 50.0;
-    let _ = tree.insert(
-        Some(root),
-        LocalNode {
-            local_bounds: Rect::new(0.0, y, 200.0, y + 40.0),
-            z_index: i as i32,
-            ..Default::default()
-        },
-    );
-}
-
-let _ = tree.commit();
-
-// Compute visible set.
-let filter = QueryFilter { visible_only: true, pickable_only: true };
-let viewport = Rect::new(0.0, 120.0, 200.0, 220.0);
-let visible: Vec<_> = tree.intersect_rect(viewport, filter).collect();
-assert!(visible.len() >= 2);
-```
-
 This crate is `no_std` and uses `alloc`.
-
-# Example
-
-```rust
-use understory_box_tree::{Tree, LocalNode, QueryFilter};
-use kurbo::{Rect, Affine, Vec2, Point};
-
-// Build a small tree.
-let mut tree = Tree::new();
-
-let root = tree.insert(
-    None,
-    LocalNode { local_bounds: Rect::new(0.0, 0.0, 200.0, 200.0), ..Default::default() },
-);
-
-let a = tree.insert(
-    Some(root),
-    LocalNode { local_bounds: Rect::new(10.0, 10.0, 60.0, 60.0), z_index: 0, ..Default::default() },
-);
-
-let b = tree.insert(
-    Some(root),
-    LocalNode { local_bounds: Rect::new(40.0, 40.0, 120.0, 120.0), z_index: 10, ..Default::default() },
-);
-
-let _damage0 = tree.commit();
-
-// Move node A to the right and compute damage.
-tree.set_local_transform(a, Affine::translate(Vec2::new(20.0, 0.0)));
-let damage = tree.commit();
-assert!(damage.union_rect().is_some());
-
-// Hit-test prefers the higher z-index (node B).
-let filter = QueryFilter { visible_only: true, pickable_only: true };
-let hit = tree.hit_test_point(Point::new(50.0, 50.0), filter).unwrap();
-assert_eq!(hit.node, b);
-```
-
-See the `basic` example in this crate for a runnable version with printed output.
 
 <!-- cargo-rdme end -->
